requirements.txt
labjack-ljm
keyboard